add root certificates to Alpine images
Fixes #96 (and the duplicate #97).

This allows Go's HTTP library to make HTTPS client connections with secure
certificate validation. Without this change, `net/http` in these images return
the error `"x509: failed to load system roots and no roots provided"`.

This adds 0.6MB to the image (as measured with docker history).