Add GOARM64 and GORISCV64 values to our `versions.json`
These are (currently) unused, so this is simply added data for completeness (as evidenced by the lack of generated `Dockerfile` changes).

This will matter more if/when we implement tip (although these values are currently the default values anyhow, so this is just being completely explicit in case those defaults ever change to match the weird way `GOARM` works with a default value on *some* platforms but a detected value on others).